fix(backend): golangci-lint CI 失敗 12件を修正
- unused: reorderProcedureRequest / reorderReservationTypeGroupRequest 削除
- unconvert: refund_service.go の不要な int64() 変換を除去
- paramTypeCombine: checkup_type_service_test.go の同型パラメータを _, _ に統一
- hugeParam: PasswordResetConfig を値渡し→ポインタ渡しに変更
- gosec G118: fire-and-forget goroutine に nolint:gosec を付与（意図的な context.Background 使用）
- gofmt: consultation/exam_type/insurance/permission_group request, cage/examination_type/inquiry_template/reservation_type_group model の末尾空行除去・整形
- goimports: password_reset_token_repository.go の import グループ順序修正（stdlib → 外部 → 内部）

+++ backend/internal/service/refund_service.go
@@ -46,7 +46,7 @@ func (s *refundService) Create(ctx context.Context, clinicID, billingID uint64, 
 	if sumErr != nil {
 		return nil, apperrors.Wrap(sumErr, "sum refunds")
 	}
-	totalAmount := int64(billing.TotalAmount)
+	totalAmount := billing.TotalAmount
 	if len(billing.Payments) > 0 {
 		totalAmount = billing.Payments[0].TotalAmount
 	}
